+++ src/main/java/com/cac/mpn/power/devices/AdvancedGeneratorTier3.java
@@ -4,9 +4,9 @@ import com.cac.mpn.power.core.IPowerDevice;
 import com.cac.mpn.power.core.SimplePowerTileEntity;
 
 public class AdvancedGeneratorTier3 extends SimplePowerTileEntity {
-    public static final int POWER_PER_SECOND = 1600;
-    public static final int POWER_PER_TICK = POWER_PER_SECOND / 20;
-    public static final int POWER_CAPACITY = 160000;
+    public static final long POWER_PER_SECOND = 1600L;
+    public static final long POWER_PER_TICK = POWER_PER_SECOND / 20L;
+    public static final long POWER_CAPACITY = 160000L;
 
     public AdvancedGeneratorTier3() {
         super(POWER_CAPACITY, IPowerDevice.DeviceType.GENERATOR);
