+++ src/main/java/com/cac/mpn/power/devices/AdvancedGeneratorTier4.java
@@ -4,9 +4,9 @@ import com.cac.mpn.power.core.IPowerDevice;
 import com.cac.mpn.power.core.SimplePowerTileEntity;
 
 public class AdvancedGeneratorTier4 extends SimplePowerTileEntity {
-    public static final int POWER_PER_SECOND = 6400;
-    public static final int POWER_PER_TICK = POWER_PER_SECOND / 20;
-    public static final int POWER_CAPACITY = 640000;
+    public static final long POWER_PER_SECOND = 6400L;
+    public static final long POWER_PER_TICK = POWER_PER_SECOND / 20L;
+    public static final long POWER_CAPACITY = 640000L;
 
     public AdvancedGeneratorTier4() {
         super(POWER_CAPACITY, IPowerDevice.DeviceType.GENERATOR);
