+++ src/main/java/com/cac/mpn/power/devices/AdvancedGeneratorTier5.java
@@ -4,9 +4,9 @@ import com.cac.mpn.power.core.IPowerDevice;
 import com.cac.mpn.power.core.SimplePowerTileEntity;
 
 public class AdvancedGeneratorTier5 extends SimplePowerTileEntity {
-    public static final int POWER_PER_SECOND = 25600;
-    public static final int POWER_PER_TICK = POWER_PER_SECOND / 20;
-    public static final int POWER_CAPACITY = 2560000;
+    public static final long POWER_PER_SECOND = 25600L;
+    public static final long POWER_PER_TICK = POWER_PER_SECOND / 20L;
+    public static final long POWER_CAPACITY = 2560000L;
 
     public AdvancedGeneratorTier5() {
         super(POWER_CAPACITY, IPowerDevice.DeviceType.GENERATOR);
