+++ src/main/java/com/cac/mpn/power/devices/AdvancedGeneratorTier6.java
@@ -4,9 +4,9 @@ import com.cac.mpn.power.core.IPowerDevice;
 import com.cac.mpn.power.core.SimplePowerTileEntity;
 
 public class AdvancedGeneratorTier6 extends SimplePowerTileEntity {
-    public static final int POWER_PER_SECOND = 102400;
-    public static final int POWER_PER_TICK = POWER_PER_SECOND / 20;
-    public static final int POWER_CAPACITY = 10240000;
+    public static final long POWER_PER_SECOND = 102400L;
+    public static final long POWER_PER_TICK = POWER_PER_SECOND / 20L;
+    public static final long POWER_CAPACITY = 10240000L;
 
     public AdvancedGeneratorTier6() {
         super(POWER_CAPACITY, IPowerDevice.DeviceType.GENERATOR);
