+++ src/main/java/com/cac/mpn/power/devices/BasicSolarGenerator.java
@@ -4,11 +4,11 @@ import com.cac.mpn.power.core.IPowerDevice;
 import com.cac.mpn.power.core.SimplePowerTileEntity;
 
 public class BasicSolarGenerator extends SimplePowerTileEntity {
-    private static final int POWER_PER_SECOND = 15;
-    private static final int POWER_PER_TICK = POWER_PER_SECOND / 20; // 0.75, 取1
-    private static final int CAPACITY = 5000;
+    private static final long POWER_PER_SECOND = 15L;
+    private static final long POWER_PER_TICK = POWER_PER_SECOND / 20L; // 0.75, 取1
+    private static final long CAPACITY = 5000L;
     private int tickCounter = 0;
-    private int tickPower = 0;
+    private long tickPower = 0L;
 
     public BasicSolarGenerator() {
         super(CAPACITY, IPowerDevice.DeviceType.GENERATOR);
@@ -20,13 +20,13 @@ public class BasicSolarGenerator extends SimplePowerTileEntity {
             tickCounter++;
             tickPower += POWER_PER_SECOND;
             if (tickCounter >= 20) {
-                setStoredPower(getStoredPower() + tickPower / 20);
+                setStoredPower(getStoredPower() + tickPower / 20L);
                 tickCounter = 0;
-                tickPower = 0;
+                tickPower = 0L;
             }
         } else {
             tickCounter = 0;
-            tickPower = 0;
+            tickPower = 0L;
         }
     }
 }
